refactor: rename utils package to utility
