+++ src/main/java/io/github/Alathra/AlathranCommands/commands/CommandAlathra.java
@@ -4,7 +4,7 @@ import io.github.Alathra.AlathranCommands.AlathranCommands;
 import dev.jorel.commandapi.CommandAPICommand;
 import dev.jorel.commandapi.CommandPermission;
 import com.github.milkdrinkers.colorparser.ColorParser;
-import io.github.Alathra.AlathranCommands.utils.ConfigUtil;
+import io.github.Alathra.AlathranCommands.utility.ConfigUtil;
 import org.jetbrains.annotations.Nullable;
 
 public class CommandAlathra implements ConfigUtil {
