+++ src/main/java/io/github/Alathra/AlathranCommands/commands/CommandTpaDeny.java
@@ -9,7 +9,7 @@ import dev.jorel.commandapi.executors.CommandArguments;
 import io.github.Alathra.AlathranCommands.data.PlayerManager;
 import io.github.Alathra.AlathranCommands.data.model.TPARequest;
 import com.github.milkdrinkers.colorparser.ColorParser;
-import io.github.Alathra.AlathranCommands.utils.TPCfg;
+import io.github.Alathra.AlathranCommands.utility.TPCfg;
 import org.bukkit.entity.Player;
 
 public class CommandTpaDeny {
