+++ src/main/java/io/github/Alathra/AlathranCommands/commands/CommandTpo.java
@@ -7,8 +7,8 @@ import dev.jorel.commandapi.CommandPermission;
 import dev.jorel.commandapi.arguments.PlayerArgument;
 import dev.jorel.commandapi.executors.CommandArguments;
 import io.github.Alathra.AlathranCommands.AlathranCommands;
-import io.github.Alathra.AlathranCommands.utils.ConfigUtil;
-import io.github.Alathra.AlathranCommands.utils.TPCfg;
+import io.github.Alathra.AlathranCommands.utility.ConfigUtil;
+import io.github.Alathra.AlathranCommands.utility.TPCfg;
 import org.bukkit.entity.Player;
 import org.bukkit.event.player.PlayerTeleportEvent;
 import org.jetbrains.annotations.Nullable;
