+++ src/main/java/io/github/Alathra/AlathranCommands/utility/Logger.java
@@ -1,4 +1,4 @@
-package io.github.Alathra.AlathranCommands.utils;
+package io.github.Alathra.AlathranCommands.utility;
 
 
 import io.github.Alathra.AlathranCommands.AlathranCommands;
